+++ test/property.test.js
@@ -1,11 +1,11 @@
 /**
  * copyright (c) 2023 nousantx
  */
-let property = {
+let AllProperty = {
   p: "padding",
   m: "margin",
   fs: "fontSize",
   w: "width",
 };
 
-export default property;
+export default AllProperty;